@@ -1,11 +1,14 @@
 module github.com/goal-web/pipeline
 
-go 1.19
+go 1.20
 
 require (
-	github.com/goal-web/container v0.1.4
-	github.com/goal-web/contracts v0.1.62
-	github.com/pkg/errors v0.8.1
+	github.com/goal-web/container v0.2.0
+	github.com/goal-web/contracts v0.2.0
+	github.com/pkg/errors v0.9.1
 )
 
-require github.com/goal-web/supports v0.1.12 // indirect
+require (
+	github.com/apex/log v1.9.0 // indirect
+	github.com/goal-web/supports v0.2.0 // indirect
+)
