@@ -5,80 +5,80 @@ import (
 )
 
 type PurePipeline struct {
-	passable interface{}
+	passable any
 
 	pipes []NextFunc
 }
 
 type PureCallback func(stack contracts.Pipe, next NextFunc) contracts.Pipe
-type NextFunc func(passable interface{}, pipe contracts.Pipe) interface{}
-type PureDestination func(passable interface{}) interface{}
+type NextFunc func(passable any, pipe contracts.Pipe) any
+type PureDestination func(passable any) any
 
 func Pure() *PurePipeline {
 	return &PurePipeline{}
 }
 
-func (this *PurePipeline) Send(passable interface{}) contracts.Pipeline {
-	this.passable = passable
-	return this
+func (pipeline *PurePipeline) Send(passable any) contracts.Pipeline {
+	pipeline.passable = passable
+	return pipeline
 }
 
-func (this *PurePipeline) SendPure(passable interface{}) *PurePipeline {
-	this.passable = passable
-	return this
+func (pipeline *PurePipeline) SendPure(passable any) *PurePipeline {
+	pipeline.passable = passable
+	return pipeline
 }
 
-func (this *PurePipeline) Through(pipes ...interface{}) contracts.Pipeline {
+func (pipeline *PurePipeline) Through(pipes ...any) contracts.Pipeline {
 	for _, item := range pipes {
 		pipe, isNextFunc := item.(NextFunc)
 		if !isNextFunc {
 			panic(PipeArgumentError)
 		}
-		this.pipes = append(this.pipes, pipe)
+		pipeline.pipes = append(pipeline.pipes, pipe)
 	}
-	return this
+	return pipeline
 }
 
-func (this *PurePipeline) ThroughPure(pipes ...NextFunc) *PurePipeline {
-	this.pipes = append(this.pipes, pipes...)
-	return this
+func (pipeline *PurePipeline) ThroughPure(pipes ...NextFunc) *PurePipeline {
+	pipeline.pipes = append(pipeline.pipes, pipes...)
+	return pipeline
 }
 
-func (this *PurePipeline) Then(destination interface{}) interface{} {
-	return this.ArrayReduce(
-		this.reversePipes(), this.carry(), this.prepareDestination(destination),
-	)(this.passable)
+func (pipeline *PurePipeline) Then(destination any) any {
+	return pipeline.ArrayReduce(
+		pipeline.reversePipes(), pipeline.carry(), pipeline.prepareDestination(destination),
+	)(pipeline.passable)
 }
 
-func (this *PurePipeline) ThenPure(destination contracts.Pipe) interface{} {
-	return this.ArrayReduce(
-		this.reversePipes(), this.carry(), destination,
-	)(this.passable)
+func (pipeline *PurePipeline) ThenPure(destination contracts.Pipe) any {
+	return pipeline.ArrayReduce(
+		pipeline.reversePipes(), pipeline.carry(), destination,
+	)(pipeline.passable)
 }
 
-func (this *PurePipeline) carry() PureCallback {
+func (pipeline *PurePipeline) carry() PureCallback {
 	return func(stack contracts.Pipe, next NextFunc) contracts.Pipe {
-		return func(passable interface{}) interface{} {
+		return func(passable any) any {
 			return next(passable, stack)
 		}
 	}
 }
 
-func (this *PurePipeline) ArrayReduce(pipes []NextFunc, callback PureCallback, current contracts.Pipe) contracts.Pipe {
+func (pipeline *PurePipeline) ArrayReduce(pipes []NextFunc, callback PureCallback, current contracts.Pipe) contracts.Pipe {
 	for _, magicalFunc := range pipes {
 		current = callback(current, magicalFunc)
 	}
 	return current
 }
 
-func (this *PurePipeline) reversePipes() []NextFunc {
-	for from, to := 0, len(this.pipes)-1; from < to; from, to = from+1, to-1 {
-		this.pipes[from], this.pipes[to] = this.pipes[to], this.pipes[from]
+func (pipeline *PurePipeline) reversePipes() []NextFunc {
+	for from, to := 0, len(pipeline.pipes)-1; from < to; from, to = from+1, to-1 {
+		pipeline.pipes[from], pipeline.pipes[to] = pipeline.pipes[to], pipeline.pipes[from]
 	}
-	return this.pipes
+	return pipeline.pipes
 }
 
-func (this *PurePipeline) prepareDestination(destination interface{}) contracts.Pipe {
+func (pipeline *PurePipeline) prepareDestination(destination any) contracts.Pipe {
 	pipe, isPipeFunc := destination.(contracts.Pipe)
 	if !isPipeFunc {
 		panic(PipeArgumentError)
